fix(git): address remaining CodeRabbit review findings
- Extract shared parsePorcelainStatus into git-status-parser.ts
- Fix subject line exceeding 50 chars by accounting for prefix length
- Change stderr from 'pipe' to 'ignore' to prevent subprocess hangs
- Add missing matcher field to Stop hook entry in hooks.json
- Add language tag to fenced code block in SKILL.md
