+++ plugins/git/hooks/hooks.json
@@ -71,6 +71,7 @@
 		],
 		"Stop": [
 			{
+				"matcher": "*",
 				"hooks": [
 					{
 						"type": "command",
